The New College Kolhapur
To provide undergraduate education that will prepare students to become Computer Professionals. To create,share and apply knowledge in computer science,including interdisciplinary areas that extend the scope of Computer Science. To educate students to be successful,ethical and effective problem-solvers and life-long learners who will contribute positively to the economic well-being of our nation.